Ls: split into separate modules
